+++ CHANGES.md
@@ -1,3 +1,13 @@
+#### [Version 0.4.0 Beta](https://github.com/Avaq/JSLite/tree/0.4.0-beta)
+_12-Sep-2013_
+
+* Added `MyClass._PARENT` to refer to the parent class.
+* Added `ClassInstance.super()` to call a method on the parent class, despite it being
+  overridden in the class that `super` is called on.
+* Added `ClassInstance.getStatic()` to find a property in the static parent chain.
+* Added `ClassInstance.callStatic()` to call a method in the static parent chain.
+* Updated the minified version.
+
 #### [Version 0.3.0 Beta](https://github.com/Avaq/JSLite/tree/0.3.0-beta)
 _03-Sep-2013_